remove flowctl-admin and add `raw combine` to `flowctl`
flowctl-admin doesn't serve much purpose anymore.
Offer the `combine` subcommand through the new client, and take
advantage of new catalog bundling routines in its implementation.
