--- crates/flowctl-admin/Cargo.toml
@@ -1,38 +1,0 @@
-[package]
-name = "flowctl-admin"
-version.workspace = true
-rust-version.workspace = true
-edition.workspace = true
-authors.workspace = true
-homepage.workspace = true
-repository.workspace = true
-license.workspace = true
-
-[dependencies]
-assemble = { path = "../assemble" }
-build = { path = "../build" }
-doc = { path = "../doc" }
-flow_cli_common = { path = "../flow_cli_common" }
-labels = { path = "../labels" }
-models = { path = "../models" }
-protocol = { path = "../protocol" }
-sources = { path = "../sources" }
-tables = { path = "../tables" }
-validation = { path = "../validation" }
-
-anyhow = { workspace = true }
-bytes = { workspace = true }
-clap = { workspace = true }
-futures = { workspace = true }
-itertools = { workspace = true }
-reqwest = { workspace = true }
-serde_json = { workspace = true }
-superslice = { workspace = true }
-tempfile = { workspace = true }
-tokio = { workspace = true }
-tracing = { workspace = true }
-url = { workspace = true }
-
-[dev-dependencies]
-assert_cmd = { workspace = true }
-tempfile = { workspace = true }
