+++ crates/flowctl/Cargo.toml
@@ -27,18 +27,19 @@ bytes = { workspace = true }
 clap = { workspace = true }
 comfy-table = { workspace = true }
 crossterm = { workspace = true }
-humantime = { workspace = true }
 dirs = { workspace = true }
 futures = { workspace = true }
+humantime = { workspace = true }
 itertools = { workspace = true }
 postgrest = { workspace = true }
 reqwest = { workspace = true }
 serde = { workspace = true }
+serde-transcode = { workspace = true }
 serde_json = { workspace = true }
 serde_yaml = { workspace = true }
-serde-transcode = { workspace = true }
 size = { workspace = true }
 superslice = { workspace = true }
+tempfile = { workspace = true }
 time = { workspace = true }
 tokio = { workspace = true }
 tokio-util = { workspace = true }
